feat : 스터디 그룹 수정에 스터디 그룹 진행 기간, 한 줄 소개 추가

+++ src/main/java/com/gamzabat/algohub/feature/studygroup/service/StudyGroupService.java
@@ -107,8 +107,15 @@ public class StudyGroupService {
 
 		if(request.name() != null && !request.name().equals(group.getName()))
 			group.editName(request.name());
+		if (request.startDate() != null)
+			group.editStartDate(request.startDate());
+		if (request.endDate() != null)
+			group.editEndDate(request.endDate());
+		if (request.introduction() != null)
+			group.editIntroduction(request.introduction());
 		if(groupImage != null){
-			imageService.deleteImage(group.getGroupImage());
+			if(group.getGroupImage() != null)
+				imageService.deleteImage(group.getGroupImage());
 			String imageUrl = imageService.saveImage(groupImage);
 			group.editGroupImage(imageUrl);
 		}
